--- pages/projects/upload/v2/index.tsx
@@ -1,68 +1,0 @@
-import styled from '@emotion/styled';
-import { useQueryClient } from '@tanstack/react-query';
-import { useRouter } from 'next/router';
-
-import { useGetMemberOfMe } from '@/api/endpoint/members/getMemberOfMe';
-import AuthRequired from '@/components/auth/AuthRequired';
-import { Confirm } from '@/components/common/Modal/Confirm';
-import useToast from '@/components/common/Toast/useToast';
-import useEventLogger from '@/components/eventLogger/hooks/useEventLogger';
-import ProjectForm from '@/components/projects/form/ProjectForm';
-import { ProjectFormType } from '@/components/projects/form/schema';
-import useCreateProjectMutation from '@/components/projects/upload/hooks/useCreateProjectMutation';
-import { getProjectListQueryKey } from '@/components/projects/upload/hooks/useGetProjectListQuery';
-import { convertToProjectData } from '@/components/projects/utils';
-import { playgroundLink } from '@/constants/links';
-import { MOBILE_MEDIA_QUERY } from '@/styles/mediaQuery';
-import { setLayout } from '@/utils/layout';
-
-const ProjectUploadPage = () => {
-  const { mutate: createProjectMutate } = useCreateProjectMutation();
-  const { data: myProfileData } = useGetMemberOfMe();
-  const router = useRouter();
-  const toast = useToast();
-  const queryClient = useQueryClient();
-  const { logSubmitEvent } = useEventLogger();
-
-  const handleSubmit = async (formData: ProjectFormType) => {
-    const notify = await Confirm({
-      title: '알림',
-      content: '프로젝트를 업로드 하시겠습니까?',
-    });
-    if (notify && myProfileData) {
-      createProjectMutate(convertToProjectData(formData, myProfileData.id)),
-        {
-          onSuccess: () => {
-            toast.show({ message: '프로젝트를 성공적으로 업로드 했어요.' });
-            router.push(playgroundLink.projectList());
-            queryClient.invalidateQueries(getProjectListQueryKey());
-            logSubmitEvent('projectUpload', {
-              writerId: String(myProfileData.id),
-            });
-          },
-        };
-    }
-  };
-
-  return (
-    <AuthRequired>
-      <Container>
-        <ProjectForm onSubmit={handleSubmit} submitButtonContent='프로젝트 등록하기' />
-      </Container>
-    </AuthRequired>
-  );
-};
-
-export default ProjectUploadPage;
-
-setLayout(ProjectUploadPage, 'header');
-
-const Container = styled.div`
-  display: flex;
-  justify-content: center;
-  margin: 98px 0;
-
-  @media ${MOBILE_MEDIA_QUERY} {
-    margin-top: 12px;
-  }
-`;
